requirements.txt
ipykernel
pandas
numpy
matplotlib
plotly
seaborn
scikit-learn
python-dotenv
pymongo
pymongo[srv]==3.12
from_root
dill
certifi
PyYAML
boto3
mypy-boto3-s3
botocore
fastapi
python-multipart
uvicorn
jinja2
imblearn
-e .